patch changes on overall business logic

+++ routes/v1/amenities.php
@@ -9,7 +9,7 @@ Route::prefix('amenity')
     ->controller(AmenityController::class)
     ->group(function () {
         Route::get('/', 'index')->name('index')->middleware('permission:properties,view');
-        Route::post('/', 'store')->name('store');
-        Route::put('/{amenity}', 'update')->name('update');
-        Route::delete('/{amenity}', 'destroy')->name('destroy');
+        Route::post('/', 'store')->name('store')->middleware('permission:properties,create');
+        Route::put('/{amenity}', 'update')->name('update')->middleware('permission:properties,update');
+        Route::delete('/{amenity}', 'destroy')->name('destroy')->middleware('permission:properties,delete');
     });
